Rename gtkgamma to gtk3gamma.
